Updated encounter flow
Fixed encounter flow by using more coroutines and awaiting results.
The PlayerStatsController now has the same terminology and methods as
the EnemyController (ModifyHealth, GetHealth) and encounters can be
started and won.

Losing an encounter has NOT been checked. It's known that the enemy
is not destroyed when the player wins and it's known that the teleport
back doesn't always work properly.